refactor: aws config 클래스 자원별 분리 - 상이한 자원별 config 클래스 분리
