--- src/main/java/com/agenticcp/core/common/config/AwsConfig.java
@@ -1,24 +1,0 @@
-package com.agenticcp.core.common.config;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import software.amazon.awssdk.regions.Region;
-import software.amazon.awssdk.services.ec2.Ec2Client;
-import software.amazon.awssdk.auth.credentials.DefaultCredentialsProvider;
-
-@Configuration
-public class AwsConfig {
-
-    @Value("${aws.region:us-east-1}")
-    private String awsRegion;
-
-    @Bean
-    public Ec2Client ec2Client() {
-        return Ec2Client.builder()
-            .region(Region.of(awsRegion))
-            .credentialsProvider(DefaultCredentialsProvider.create())
-            .build();
-    }
-}
